Switch boundedness of job and result queue
The problem with the bounded job queue is that processing a result may
produce more jobs, leading to a deadlock.

This change is also obvious if you think about it in terms of
backpressure: we need to restrict the flow of incoming data to a level
that lets the threads that process the data catch up.

Discovered by setting the sizes of all bounded queues to 1.